rename clean up on utilities rework of serversettings class injection into the auth
