+++ Server/Utilities/TryResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Server.Utiilites
+﻿namespace Server.Utilities
 {
     public class TryResult<T>
     {
